Remove 'slip' from requirements.txt (#731)
'slip' is a firewalld dependency. In CentOS repository there's
a slip 0.4.0 but in pip repository there's no such version.

If firewalld is not installed then pip tries to fetch the 0.4.0
version from its repo which fails with 'Could not find a version
that satisfies the requirement slip==0.4.0' error.

As slip is only needed for firewalld we can remove it from our
requirements.