Update configMap.yml
The default poll mode driver of virtual functions for Intel Corporation I350 NIC is igbvf. In order to allocate sriov net devices bound to igbvf driver, they need to be added to the intel_sriov_netdevice resource pool by the sriov device plugin. The above changes allow the VFs to be allocated to the pods.